Refactor code. Update the folder from util to noteutil.
