+++ backend/note/api/v1/transport/rest/delete_handler_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"noteapp/note"
-	"noteapp/note/util/copyutil"
+	"noteapp/note/noteutil/copyutil"
 )
 
 func (s *HandlerTestSuite) TestDelete() {
