+++ backend/note/api/v1/transport/rest/update_handler_test.go
@@ -9,7 +9,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"noteapp/note"
-	"noteapp/note/util/copyutil"
+	"noteapp/note/noteutil/copyutil"
 	"noteapp/pkg/ptrconv"
 	"noteapp/pkg/timestamp"
 )
