+++ backend/note/noteutil/copyutil/copyutil.go
