+++ backend/note/store/file/store_test.go
@@ -19,14 +19,16 @@ type FileStoreTestSuite struct {
 }
 
 func (s *FileStoreTestSuite) SetupTest() {
-
-	fs := afero.NewOsFs()
+	fs := afero.NewMemMapFs()
 	file, err := fs.OpenFile("./test_note.pb", os.O_CREATE|os.O_RDWR|os.O_TRUNC, 0666)
 	require.NoError(s.T(), err)
 
 	s.SetStore(Must(newStore(file)))
 }
 
-func (s *FileStoreTestSuite) Test() {
-	s.TestSuite.TestDelete()
+func (s *FileStoreTestSuite) TestInsert() {
+	s.TestSuite.TestInsert()
+
+	// Extend the test.
+
 }
